Update TDD stop RX
When doing TDD stop procedure should do not wait for rx_allowed_on event to pass.
Explanation: In cell search mode (DDD..D), the rx_allowed_on was far away in the past.
In particular circumstances that could be considered a future event and waiting for event to pass will result in a wrong behavior

Signed-off-by: Florin Chiculita <florinlaurentiu.chiculita@nxp.com>